Greedy times update

+++ L01.Working-With-Abstraction/Problems-Solutions/GreedyTimes/Models/Preciouses/Cash.cs
@@ -2,7 +2,7 @@
 {
     public class Cash : Precious
     {
-        public Cash(string preciousType, int qtty) 
+        public Cash(string preciousType, long qtty) 
             : base(preciousType, qtty)
         {
         }
