+++ L01.Working-With-Abstraction/Problems-Solutions/GreedyTimes/Models/Preciouses/Gem.cs
@@ -2,7 +2,7 @@
 {
     public class Gem : Precious
     {
-        public Gem(string preciousType, int qtty) 
+        public Gem(string preciousType, long qtty) 
             : base(preciousType, qtty)
         {
         }
